timer: Fix counter wrap around issue
The KONA timer documentation states that if a programmed match
value is less than 65K counts behind the counter value, the
timer hardware treats it as an expiry and immediately generates
a match. This theory mostly holds except for the corner case when
the counter value wraps. If the counter value wraps before the
state machine has completed syncing/enabling the match logic,
the match value is treated as an expiry far in the future. In
that case the match is triggered only when the counter value
reaches or crosses the programmed match value, which is ~71
minutes for a 1MHz counter.

The counter wrap around issue is addressed as follows:

For timer channels >= 1MHz, if the programmed match value is
within 65K of the counter wrap around value, first allow the
channel to sync/enable the match logic. If the counter wraps
by the time matching logic starts, we reprogram the match value
to be the same as the counter value, to generate a match
immediately.

Change-Id: If862995c725dbba471699cce5f500647067db2af
Signed-off-by: Danesh Petigara <dpetigara@broadcom.com>
Reviewed-on: http://mps-gerrit.broadcom.com/61298
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Reviewed-by: Commit Description Check <commit_desc_noreply@broadcom.com>
Reviewed-by: Checkpatch Status <checkpatch_status_noreply@broadcom.com>
Reviewed-by: Edwin Chan <echan@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>